Added signal handlers to create assets as needed
