+++ items/apps.py
@@ -3,3 +3,6 @@ from django.apps import AppConfig
 
 class ItemsConfig(AppConfig):
     name = 'items'
+
+    def ready(self):
+        import items.signals.handlers
