+++ items/models/item_models.py
@@ -1,8 +1,4 @@
 from django.db import models
-from django.db.models.signals import post_save
-from django.dispatch import receiver
-
-from items.logic.email_logic import send_email_with_item_under_threshold
 
 ITEM_HEADERS = ['name', 'quantity', 'description', 'model_number', 'tags', 'minimum_stock', 'track_minimum_stock']
 
@@ -19,12 +15,6 @@ class Item(models.Model):
     def __str__(self):
         item_string = "{name} : {quantity}".format
         return item_string(name=self.name, quantity=self.quantity)
-
-
-@receiver(post_save, sender=Item)
-def minimum_quantity_callback(instance, **kwargs):
-    if instance.quantity < instance.minimum_stock and instance.track_minimum_stock:
-        send_email_with_item_under_threshold(instance)
 
 
 class Tag(models.Model):
